mongodb_dashboard/requirements.txt
flask
pymongo
pandas
plotly
scikit-learn
folium
python-dotenv
matplotlib
routes